feat: CoinPriceHistory with TimeRangeSelector

+++ backend/tracker/src/main/java/com/crypto/tracker/repos/CoinRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.crypto.tracker.model.Coin;
 
-public interface CoinRepository extends JpaRepository<Coin, Long> {
+public interface CoinRepository extends JpaRepository<Coin, Integer> {
     // Zusätzliche Query-Methoden können hier ergänzt werden
-    Optional<Coin> findById(Long id);
+    Optional<Coin> findById(Integer id);
 }
